refactor evaluation policy logic to evaluation
