+++ src/main/java/depromeet/lessonfour/server/report/domain/vo/activity/FoodEvaluation.java
@@ -1,24 +1,52 @@
-package depromeet.lessonfour.server.report.domain.policy;
+package depromeet.lessonfour.server.report.domain.vo.activity;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.springframework.stereotype.Component;
-
 import depromeet.lessonfour.server.activityrecord.domain.entity.FoodRecord;
 import depromeet.lessonfour.server.activityrecord.domain.vo.MealTime;
-import depromeet.lessonfour.server.report.domain.vo.DayType;
-import depromeet.lessonfour.server.report.domain.vo.FoodEvaluation;
+import lombok.Getter;
 
-@Component
-public class FoodEvaluationPolicy {
+@Getter
+public class FoodEvaluation {
 
   private static final int DEFAULT_DANGEROUS_THRESHOLD = 80;
 
-  public FoodEvaluation calculate(List<FoodRecord> records, DayType dayType) {
-    if (records == null || records.isEmpty()) {
+  private final DayType dayType;
+  private final boolean dangerous;
+  private final Map<MealTime, List<String>> foodsByMealTime;
+  private final Set<MealTime> dangerousMealTimes;
+
+  public FoodEvaluation(
+      boolean dangerous,
+      Map<MealTime, List<String>> foodsByMealTime,
+      DayType dayType,
+      Set<MealTime> dangerousMealTimes) {
+    this.dangerous = dangerous;
+    this.dayType = Objects.requireNonNull(dayType, "dayType must not be null");
+
+    Map<MealTime, List<String>> source =
+        (foodsByMealTime == null) ? Collections.emptyMap() : foodsByMealTime;
+    Map<MealTime, List<String>> copy = new HashMap<>(source.size());
+
+    for (Map.Entry<MealTime, List<String>> e : source.entrySet()) {
+      List<String> list =
+          (e.getValue() == null) ? Collections.emptyList() : List.copyOf(e.getValue());
+      copy.put(e.getKey(), list);
+    }
+
+    this.foodsByMealTime = Collections.unmodifiableMap(copy);
+    this.dangerousMealTimes =
+        dangerousMealTimes == null ? Collections.emptySet() : Set.copyOf(dangerousMealTimes);
+  }
+
+  public static FoodEvaluation calculate(List<FoodRecord> records, DayType dayType) {
+    if (records.isEmpty()) {
       return new FoodEvaluation(false, Map.of(), dayType, Set.of());
     }
 
@@ -52,5 +80,9 @@ public class FoodEvaluationPolicy {
             .collect(Collectors.toSet());
 
     return new FoodEvaluation(dangerous, foodsByMealTime, dayType, dangerousMealTimes);
+  }
+
+  public boolean isMealDangerous(MealTime mealTime) {
+    return dangerousMealTimes.contains(mealTime);
   }
 }
